Add classes to enable the configuration of the Oak repository from a properties file. This is an alternative to the OSGi configuration in Oak that has the drawback to depend on a lot of dependencies on both OSGi and Felix (an implementation of OSGi).
The creation of the Oak repository is based upon the RepositoryFactory
that is loaded from the Java Service Provider as expected in the JCR
specifications. Silverpeas provides an implementaton of this
RepositoryFactory interface. The creation of the Oak repository is
driven by the configuration loaded from a properties file whose the path
has to be provided by the system property jcr.conf. It is the goal of
the RepositoryFactory to create and configure the Oak repository from
the loaed configuration properties whereas it is the goal of the
RepositoryProvider to look for an available RepositoryFactory that knows
how to treat the peculiar Silverpeas parameters for the JCR. Any use of
the repository should ask for it to the RepositoryProvider that is a
singleton.

Add a JAAS LoginModule to perform the authentication of a user by its
API token.

Add some unit tests to validate the creation of an Oak repository. For
the tests, the repository is based upon a memory storage for the data.