Update version of Oak to 1.48.0.
SilverpeasContentRepository is renamed to SilverpeasRepository which is
created by SilverpeasRepositoryFactory. This is for have coherence
with the names between the object and his factory.

The NodeStore in Oak requires to be closed/disposed at the ends of the
application using the JCR in order to free some allocated resources (for
example, the lock on the storage with the SegmentNodeStore.

For doing, introduce the OakRepository and OakRepositoryConnection
classes dedicated to wrap the Oak's repository implementation and to
handle both the creation and the close of the repository (and hence of
the NodeStore used behind). A new method is added to the interface
SilverpeasRepositoryFactory: closeRepository; it is the symetric
operation of getting a repository in the RepositoryFactory interface.
OakRepository extends SilverpeasRepository.

Refine the configuration for Segment Storage in Oak. The property
segment.repository.home is renamed to segment.repository and it
refers now the path of the directory into which the content of the JCR
is stored. By default, it's named segmentstore (see doc of Oak) and it
is located as a subdirectory of the JCR home directory.